Cleanup finished and merged with development (#6)
* Restructured every page and utility into seperate files

* Finished restructuring and cleanup, 1 error to be fixed

* fixed Hive not opening the box

* Fixed json parser keys
